llm_playground/requirements.txt
Flask>=1.1.4
Jinja2>=2.10.1,<=3.0.2
MarkupSafe==2.0.1
python-dotenv==0.19.1
openai==0.28.0
